+++ app/EmotionRootStyleRegistry.tsx
@@ -1,0 +1,35 @@
+'use client';
+
+import { CacheProvider } from '@emotion/react';
+import createCache from '@emotion/cache';
+import { useServerInsertedHTML } from 'next/navigation';
+import { useState } from 'react';
+
+import ChakraUIWrapper from './ChakraUIWrapper';
+
+export default function EmotionRootStyleRegistry({ children }: { children: JSX.Element }) {
+    const [cache] = useState(() => {
+        const cache = createCache({ key: 'css' });
+        cache.compat = true;
+        return cache;
+    });
+
+    useServerInsertedHTML(() => {
+        return (
+            <style
+                data-emotion={`${cache.key} ${Object.keys(cache.inserted).join(' ')}`}
+                dangerouslySetInnerHTML={{
+                    __html: Object.values(cache.inserted).join(' '),
+                }}
+            />
+        );
+    });
+
+    return (
+        <CacheProvider value={cache}>
+            <ChakraUIWrapper>
+                <div>{children}</div>
+            </ChakraUIWrapper>
+        </CacheProvider>
+    );
+}
