+++ app/side-by-side/[userName]/[repoName]/side-by-side.module.css
@@ -1,0 +1,7 @@
+.sideBySide {
+    display: flex;
+    flex-direction: row;
+}
+.side {
+    width: 50vw;
+}
