+++ app/waterfall/[userName]/[repoName]/page.tsx
@@ -1,0 +1,28 @@
+'use client';
+import React, { Suspense } from 'react';
+
+import { ContentLoadingFallback } from '../../../../components/ContentLoadingFallback';
+import RichErrorBoundary from '../../../../components/RichErrorBoundary';
+import InternalLink from '../../../../components/InternalLink';
+
+import { UserRepoWaterfall } from '../../../../container/LazyUserRepo';
+import { UserRepoFromUrlProvider } from '../../../../components/useUserRepoFromRoute';
+
+export default function WaterfallPage() {
+    return (
+        <UserRepoFromUrlProvider>
+            <InternalLink href={'/waterfall'}>back to shortcut list</InternalLink>
+            <WaterfallMain />
+        </UserRepoFromUrlProvider>
+    );
+}
+
+export function WaterfallMain() {
+    return (
+        <RichErrorBoundary>
+            <Suspense fallback={<ContentLoadingFallback />}>
+                <UserRepoWaterfall />
+            </Suspense>
+        </RichErrorBoundary>
+    );
+}
