+++ components/InternalLink.tsx
@@ -1,3 +1,4 @@
+'use client';
 import NextLink, { type LinkProps as NextLinkProps } from 'next/link';
 import { chakra } from '@chakra-ui/react';
