+++ components/PullRequestInfo.tsx
@@ -1,3 +1,4 @@
+'use client';
 import React, { useEffect, useState } from 'react';
 import { Button, Icon, Link, VStack } from '@chakra-ui/react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
@@ -8,7 +9,7 @@ import { Maybe } from '../restinpeace/types';
 import { fetchRepoPullRequestsAssociatedWithCommit } from '../restinpeace/github';
 import { createResource } from '../cache/reactCache';
 import { useUserRepo } from './useUserRepoFromRoute';
-import { useDoMergePR } from '../pages/waterfall/[userName]/[repoName]';
+//import { useDoMergePR } from '../pages/waterfall/[userName]/[repoName]';
 
 export type PullRequestInfoProps = {
     pullRequest?: PullRequestData;
@@ -28,7 +29,7 @@ const getPR = createResource(
 
 export default function PullRequestInfo({ pullRequest, sha }: PullRequestInfoProps) {
     const { userName, repoName } = useUserRepo();
-    const doMergePR = useDoMergePR({ userName, repoName });
+    //const doMergePR = useDoMergePR({ userName, repoName });
 
     const [mergeRequest, setMergeRequest] = useState<Promise<unknown>>();
     const [isMerged, setIsMerged] = useState<boolean>(false);
@@ -59,7 +60,8 @@ export default function PullRequestInfo({ pullRequest, sha }: PullRequestInfoPro
             <Link href={html_url ?? url ?? ''} title={title ?? ''} rel="noopener noreferrer nofollow">
                 #{number}
             </Link>
-            {doMergePR && !isMerged && (
+            {/*
+                doMergePR && !isMerged && (
                 <Button
                     ml={1}
                     size="xs"
@@ -68,7 +70,8 @@ export default function PullRequestInfo({ pullRequest, sha }: PullRequestInfoPro
                     disabled={!!mergeRequest}>
                     Rebase&Merge
                 </Button>
-            )}
+            )
+            */}
             {!!error && (
                 <Icon>
                     <FontAwesomeIcon icon={faExclamationTriangle as IconProp} size="1x" />
