+++ components/Spinner.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import React from 'react';
 
 import { CircularProgress, CircularProgressProps } from '@chakra-ui/react';
