+++ components/isInBrowserEnv.ts
@@ -1,0 +1,3 @@
+export function isInBrowserEnv() {
+  return typeof window !== 'undefined';
+}
