+++ container/UserRepo.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import React from 'react';
 import { Alert, AlertIcon, Flex } from '@chakra-ui/react';
