--- pages/side-by-side/[userName]/side-by-side.module.css
@@ -1,8 +1,0 @@
-.sideBySide {
-    display: flex;
-    flex-direction: row;
-    _width: 100vw;
-}
-.side {
-    width: 50vw;
-}
